Verificar Excel
7-3-2015

+++ PrimerProjectoWeb/src/java/com/luis/clases/Excel.java
@@ -124,7 +124,7 @@ public class Excel {
 		switch (celda.getCellType()) {
 
 			case 0:
-				valor = "" + celda.getNumericCellValue();
+				valor = "" + (int)celda.getNumericCellValue();
 				break;
 			case 1:
 				valor = "" + celda.getStringCellValue();
